Add couple-sum, spenddown, and resource boundary tests for the Medicaid ABD unit
Address review: isolate the marital-unit sum with a Missouri couple each
under the individual standard but over the couple standard (income,
209(b) spenddown, and resources), add a blind couple spenddown case, make
the per-child deeming split a float array, and clarify the April 1 vs
January guideline timing and the SSI federal benefit rate labels.

+++ policyengine_us/variables/gov/hhs/medicaid/income/deemed/from_ineligible_parent/medicaid_optional_senior_or_disabled_unearned_income_deemed_from_ineligible_parent.py
@@ -85,7 +85,7 @@ class medicaid_optional_senior_or_disabled_unearned_income_deemed_from_ineligibl
         )
 
         count_eligible_children = tax_unit.sum(eligible_child)
-        income = np.zeros_like(count_eligible_children)
+        income = np.zeros_like(count_eligible_children, dtype=float)
         mask = count_eligible_children > 0
         income[mask] = net_parental_deemed_income[mask] / count_eligible_children[mask]
         return income
